turn on dirtrack mode at startup
When using shell mode, dirtrack mode allows find-file and similar
commands to start at the PWD of the shell, rather than the directory
in which the shell was started. Every time I have ever called
find-file while in a shell, I have wanted the former behavior (current
PWD), and not the latter. This causes that.