add new binding for inactive time stamp
Not only is this binding purposefully similar to the one for an active
timestamp, it also avoids the problem I had where Flycheck's default
C-c ! prefix was overriding commands in org's mode map. (This is of
course a result of minor mode bindings overriding major mode
bindings.) In any case I always thought that C-c ! was a stupid place
to put the inactive-time-stamp binding, so having my hand forced by
Flycheck's override turned out to be a good thing. Now I have a useful
binding that is easy to remember because it's similar to the
active-timestamp binding.