set aw-minibuffer-flag to t
Normally the current state for ace-window is displayed in the modeline
for the buffer which was current when you called ace-window. That
displayed string can easily get cut off when the window displaying the
buffer is small. This flag avoids that problem by displaying the
ace-window state (and selected action, if not the default) in the
minibuffer. It usually works, though I've had it fail to fire at least once.