abbrev nonsense
This way of maintaing an abbrev file is really dumb, because it
results in all kinds of line noise in commits. Every time you use an
abbrev it updates a counter; I don't care about that at all, yet it
means yet another line in the eventual Big Giant Abbrev File Commit
which communicates nothing useful.

I'll have to find a way to just set up my abbrevs and have Emacs not
touch them with these silly incrementing counters or any other
nonsense. I'll figure it out eventually, but it's not high-priority.
Just a teeny thorn in the side.