improve function which visits my main org file
Before, all it did was wrap a call to find-file.

Now, if there's already a window (in current frame) with a buffer
visiting that file, it just selects that window instead.

Much nicer when working with lots of windows, since my org file is
usually already open and visible. One less window-management task to
do manually sounds good to me.